move naics functions from `validation.py` to `naics.py`
